[GH-#35] removed aecore.block.* and aecore.txs.tx
