+++ apps/aecore/test/aecore_hashcash_test.exs
@@ -3,11 +3,11 @@ defmodule HashcashTest do
   doctest Aecore.Pow.Hashcash
 
   alias Aecore.Pow.Hashcash
-  alias Aecore.Block.Genesis
+  alias Aecore.Structures.Block
 
   @tag timeout: 10000000
   test "successfull test" do
-    assert {:ok, mined_header} = Hashcash.generate(Genesis.genesis_block.header)
+    assert {:ok, mined_header} = Hashcash.generate(Block.genesis_block.header)
     assert :true = Hashcash.verify(mined_header)
   end
 end
