+++ apps/aecore/test/aecore_tx_test.exs
@@ -4,9 +4,7 @@ defmodule AecoreTxTest do
   """
 
   use ExUnit.Case
-  doctest Aecore.Txs.Tx
 
-  alias Aecore.Txs.Tx, as: Tx
   alias Aecore.Keys.Worker, as: Keys
   alias Aecore.Structures.SignedTx
 
@@ -16,9 +14,11 @@ defmodule AecoreTxTest do
   end
 
   test "create and verify a signed tx" do
-    tx = Tx.create("to_account",5)
+    {:ok, to_account} = Keys.pubkey()
+    tx = SignedTx.create(to_account, 5)
+
     assert %SignedTx{} = tx
-    assert :true = Tx.verify(tx)
+    assert :true = SignedTx.verify(tx)
   end
 
 end
